Bug fixes
* Make sure node->pri will not go beyond LONG_MAX / LONG_MIN
* empty_list() did leave a corrupted struct list_s behind

+++ src/nodelist.h
@@ -39,7 +39,7 @@
 typedef struct kw_node_s {
         struct kw_node_s *succ;
         struct kw_node_s *pred;
-        int pri;
+        long pri;
         char *nodename;
 } kw_node_t;
